src/app/requirements.txt
Django
psycopg2-binary
gunicorn
Pillow
asgiref
django-heroku
psycopg2
sqlparse
whitenoise
dj-database-url
django-braces
